requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.23.0
open3d>=0.17.0
pymeshlab>=2022.2
gdown>=4.7.0
